==> code/02_cleaning.ipynb <==
#### Step 2: Cleaning

import pandas as pd

To begin the cleaning process, I'll take a look at the data at a high level to see how many null values there might be and if there are any surprises about how the data types are being stored.

data = pd.read_csv('../data/posts.csv')

data.info()

Only about 60% of the samples taken have selftext, but otherwise the data seems to be intact. The lack of selftext won't matter anyways as I've decided to just look at the titles of posts.

The selftext will be helpful, however, in determining if a post has been removed. Because these removed posts may be off-topic posts removed by moderators, duplicates, or otherwise non-relevant, I will exclude them from the analysis. 

data = data[data['selftext'] != '[removed]']

With the removed posts removed, I can drop columns to only leave the features that we'll be using in the model: the target ('subreddit') and the title of the posts. 

data = data.drop(columns=['selftext', 'created_utc'])

Once we're down to only subreddits and titles, I want to drop any duplicate post titles that still may be in the set, as I don't want something like a recurring thread to dominate the predictions. 

data = data.drop_duplicates()

With our basic cleaning out of the way, we can check to see how many samples are still left:

data.info()

In total it looks like we dropped 393 posts, or just under 8% of the sample. This should still leave us with enough posts to create a reasonable model, as long as our class split didn't become too skewed.

data['subreddit'].value_counts()

Looks like r/tea lost a few more posts than r/Coffee did, but the two classes are still approximately equal, and shouldn't cause any imbalance issues.

With cleaning complete, I'll save the new dataframe for use in EDA and modeling:

data.to_csv('../data/cleaned_posts.csv', index=False)
