==> code/04_modeling.ipynb <==
#### Step 3: Modeling

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from nltk import word_tokenize          
from nltk.stem import WordNetLemmatizer

from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier

data = pd.read_csv('../data/cleaned_posts.csv')

X = data['title']
y = data['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=1120)

Before starting to evaluate any classification models, it's important to determine a baseline of accuracy for comparing results. This null model will use the most commonly occuring target value in the training data as a prediction for every value in the test data. 

#### Step 3A: Null Model

y_train.value_counts(normalize=True)

y_test.value_counts(normalize=True)

Guessing 'Coffee' for every sample gives us a null accuracy of 50.3%, only slightly off from the 50% we would have expected had all samples been kept and the classes were exactly even.

#### Step 3B: Natural Language Processing & Vectorization

The first step in our modeling process will be to transform our text. This will include any preprocessing in the form of tokenization, stemming, or lemmatization, the actual transformer that we'll use to vectorize the tokens, and the hyperparameters that we'll feed into that transformer, like n-gram size, the minimum and maximum number of samples a token must occur in to be included, and the number of tokens we ultimately want to see used as features.

There are two main vectorizers we've been using to consider:
- CountVectorizer, which gives a simple count of tokens as matrix values
- TfidfVectorizer (TFDIF: 'term frequency'-'inverse document frequency'), which creates a matrix which assigns value based on how often a term occurs in the target class vs how often it occurs in the sample as a whole

Both vectorizers will require a tokenizer to be specified to break up the text into words or phrases as well as process things like punctuation. This could be a simple tokenizer that just grabs the words as they appear in the sample text, or a tokenizer that includes lemmatization, the process of attempting to reduce a word to a base form in order to count all the forms a word might take as the same feature (i.e. ideally 'run,' 'ran,' and 'running' would all become 'run'). 

For this model, I will be testing the default basic tokenizer as well as a custom tokenizer that incorporates the WordNetLemmatizer. 

Further work could be done to specify custom regular expressions ('regex') for tokenization or a more powerful NLP library like spaCy could be used, but prior testing has shown that neither one will provide a significant improvement to this model to justify their complexities.

# this custom wordnetlemmatizer code has been taken from:
# https://stackoverflow.com/questions/47423854/sklearn-adding-lemmatizer-to-countvectorizer
# and is similar to code found in lesson 5.04 NLP II

class LemmaTokenizer(object):
    def __init__(self):
        self.wnl = WordNetLemmatizer()
    def __call__(self, articles):
        return [self.wnl.lemmatize(t) for t in word_tokenize(articles)]

To test what kind of vectorizer to use, I'll use GridSearchCV to test out different combinations of vectorizers, tokenizers, and hyperparameters. For comparison, they will all use a naive Bayes classifier. 

vector_pipe = Pipeline([
    ('vect', CountVectorizer()),
    ('bayes', MultinomialNB())
])

vector_params = {
    'vect' : [CountVectorizer(), TfidfVectorizer()],
    'vect__tokenizer' : [None, LemmaTokenizer()],
    'vect__stop_words' : [None, 'english'],
    'vect__ngram_range' : [(1, 1), (1, 2), (2, 2)],
    'vect__min_df' : [1, 2, 5],
    'vect__max_features' : [None, 1000, 2500]
}

vector_grid = GridSearchCV(vector_pipe, param_grid=vector_params, n_jobs=-1)

Instead of rerunning the timely search several times, I saved the results of the cell below before commenting out the code. 

# vector_grid.fit(X_train, y_train)
# vector_search = pd.DataFrame(vector_grid.cv_results_).sort_values(by='mean_test_score', ascending=False)
# vector_search.to_csv('../grid_search/vectorizers.csv', index=False)

I can then load the saved results:

vector_search = pd.read_csv('../grid_search/vectorizers.csv')

vector_search.head()

The results show that the top performing models used the CountVectorizer with minimum occurences of 1 or 2, monograms and bigrams, stop words removed, lemmatization with the custom LemmaTokenizer, and no maximum features.

I'll use the second best performing vectorization model moving forward, as limiting the features to occurring in a minimum of 2 samples will significantly reduce the number of features being used in calculations without significantly impacting accuracy.

cvect = CountVectorizer(ngram_range=(1, 2), stop_words='english', min_df=2, tokenizer=LemmaTokenizer())

#### Step 3C: Classification Model

For the classification itself, I'll be testing several different models as well as fine tuning their respective hyperparameters. I'll begin with three singular models: a naive Bayes classifier, logistic regression, and a support vector classifier.

bayes_pipe = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('bayes', MultinomialNB())
])

bayes_params = {
    'bayes__alpha' : [0.001, 1, 100, 100_000, 1_000_000]
}

bayes_grid = GridSearchCV(bayes_pipe, bayes_params, n_jobs=-1)

Like with the vectorizers above, I'll be saving and loading the results of the grid searches to keep from running the searches more than necessary. 

# bayes_grid.fit(X_train, y_train)
# bayes_search = pd.DataFrame(bayes_grid.cv_results_).sort_values(by='mean_test_score', ascending=False)
# bayes_search.to_csv('../grid_search/naivebayes.csv', index=False)

bayes_search = pd.read_csv('../grid_search/naivebayes.csv')

bayes_search.head()

Here we can again see that the large alpha parameter of 100,000 is the best performing on the dataset, with a mean score of .895. We can also check how it performs on the full training set and the test set:

bayes_pipe = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('bayes', MultinomialNB(alpha=100_000))
])

bayes_pipe.fit(X_train, y_train)

bayes_pipe.score(X_train, y_train)

bayes_pipe.score(X_test, y_test)

There's a fair gap between the train and test scores, indicating that the model is overfitting, but this tends to be the norm for a naive Bayes classifier and isn't surprising for data that could see such large differences like vocabulary choice. 

I'll search over logistic regression next:

lr_pipe = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('lr', LogisticRegression(solver='saga'))
])

lr_params= {
    'lr__penalty' : ['l1', 'l2'],
    'lr__C' : [0.001, 0.01, 0.1, 1, 10, 100, 1000],
    'lr__max_iter' : [100, 1000]
}

lr_grid = GridSearchCV(lr_pipe, param_grid=lr_params, n_jobs=-1)

# lr_grid.fit(X_train, y_train)
# lr_search = pd.DataFrame(lr_grid.cv_results_).sort_values(by='mean_test_score', ascending=False)
# lr_search.to_csv('../grid_search/logisticregression.csv', index=False)

lr_search = pd.read_csv('../grid_search/logisticregression.csv')

lr_search.head()

Here we see our highest performance with an l1 penalty, a C value of 0.1, and a maximum of 100 iterations.

The highest mean test score was .915, and we can check performance on the full training set and the test set next: 

lr_pipe1 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('lr', LogisticRegression(penalty='l1', C=.1, max_iter=100, solver='saga'))
])

lr_pipe1.fit(X_train, y_train)

lr_pipe1.score(X_train, y_train)

lr_pipe1.score(X_test, y_test)

We can see that the logistic regression outperforms the naive Bayes on the train data and the test data, but also shows signs of overfit.

With that in mind, we can check the scores with some smaller C values, corresponding with stronger regularization. 

lr_pipe2 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('lr', LogisticRegression(solver='saga', penalty='l1', max_iter=100, C=0.05))
])

lr_pipe2.fit(X_train, y_train)

lr_pipe2.score(X_train, y_train)

lr_pipe2.score(X_test, y_test)

Through some trial and error, it seems like that without majorly sacrificing train and test accuracy in the name of curbing overfit, the best parameters are an l1 penalty with a C value of 0.05.

The next model I'll try is the support vector classifier.

svc_pipe = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('svc', SVC())
])

svc_params= {
    'svc__kernel' : ['rbf', 'sigmoid', 'poly'],
    'svc__C' : [0.001, 0.01, 0.1, 1, 10, 100, 1000],
}

svc_grid = GridSearchCV(svc_pipe, param_grid=svc_params, n_jobs=-1)

# svc_grid.fit(X_train, y_train)
# svc_search = pd.DataFrame(svc_grid.cv_results_).sort_values(by='mean_test_score', ascending=False)
# svc_search.to_csv('../grid_search/svc.csv', index=False)

svc_search = pd.read_csv('../grid_search/svc.csv')

svc_search.head()

svc_pipe1 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('svc', SVC(kernel='sigmoid'))
])

svc_pipe1.fit(X_train, y_train)

svc_pipe1.score(X_train, y_train)

svc_pipe1.score(X_test, y_test)

The best model appears to be the sigmoid kernel with a C value of 1. Trial of different C values was not able to significantly reduce the gap between the train and test scores or improve the test accuracy. We can also compare the gap to the gaussian kernel:

svc_pipe2 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('svc', SVC(C=1))
])

svc_pipe2.fit(X_train, y_train)

svc_pipe2.score(X_train, y_train)

svc_pipe2.score(X_test, y_test)

Across different C values, the gaussian kernel performed worse on the test data than the sigmoid kernel. 

Next I'll look at an ensembling model, the random forest classifier.

rfc_pipe = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('rfc', RandomForestClassifier(n_jobs=-1))
])

rfc_params= {
    'rfc__criterion' : ['gini', 'entropy'],
    'rfc__n_estimators' : [10, 100, 1000],
    'rfc__max_depth' : [3, 5, 10, None]
}

rfc_grid = GridSearchCV(rfc_pipe, param_grid=rfc_params)

# rfc_grid.fit(X_train, y_train)
# rfc_search = pd.DataFrame(rfc_grid.cv_results_).sort_values(by='mean_test_score', ascending=False)
# rfc_search.to_csv('./grid_search/randomforest.csv', index=False)

rf_search = pd.read_csv('../grid_search/randomforest.csv')

rf_search.head()

The best parameters were found to be gini criterion, 1000 estimators, and no limit on depth. We can see the performance on train and test sets below.

rfc_pipe1 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('rfc', RandomForestClassifier(criterion='gini', n_estimators=1000, n_jobs=-1))
])

rfc_pipe1.fit(X_train, y_train)

rfc_pipe1.score(X_train, y_train)

rfc_pipe1.score(X_test, y_test)

The training score is very high, indicating there's definitely overfit happening. We can test to see if other parameters avoided overfit any better or were able to increase the accuracy on the test set:

rfc_pipe2 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('rfc', RandomForestClassifier(criterion='entropy', n_estimators=1000, n_jobs=-1))
])

rfc_pipe2.fit(X_train, y_train)

rfc_pipe2.score(X_train, y_train)

rfc_pipe2.score(X_test, y_test)

Entropy as a criterion seems to score slightly better on the test set than gini.  

rfc_pipe3 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('rfc', RandomForestClassifier(criterion='gini', n_estimators=100, n_jobs=-1))
])

rfc_pipe3.fit(X_train, y_train)

rfc_pipe3.score(X_train, y_train)

rfc_pipe3.score(X_test, y_test)

Seems like the higher number of estimators didn't actually add much to the model.

rfc_pipe4 = Pipeline([
    ('cvect', cvect),
    ('sscaler', StandardScaler(with_mean=False)),
    ('rfc', RandomForestClassifier(criterion='gini', n_estimators=100, max_depth=75, n_jobs=-1))
])

rfc_pipe4.fit(X_train, y_train)

rfc_pipe4.score(X_train, y_train)

rfc_pipe4.score(X_test, y_test)

No matter the parameters, the gap between the train score and the test score remains between about 4 and 8 points, but a max depth of 75 gives us the highest test set score of 90.9.
